modules.rst put back in

+++ doc/source/conf.py
@@ -342,6 +342,6 @@ if os.environ.get('READTHEDOCS', None) == 'True':
 
     for f in os.listdir("."):
         if (os.path.isfile(f) and f.endswith(".rst")
-                and f != "index.rst"):
+                and f != "index.rst" and f != "modules.rst"):
             os.remove(f)
     apidoc.main([None, '-o', "../../spinn_machine", "."])
